Add back CatcherArea to simplify layout
